fix: compile errors for nix 0.29 and libseccomp 0.3 APIs
- Fix duplicate create_dir_all import in rootfs.rs
- Update mknod calls to use nix::sys::stat imports
- Fix unshare import from nix::sched
- Add enter_container_namespaces import in container.rs
- Update libseccomp API: ScmpSyscall::from_name, ScmpFilterContext::new_filter
- Fix ScmpCompareOp variant names (Equal, NotEqual, Greater, GreaterEqual, Less, LessOrEqual, MaskedEqual)
- Fix PR_SET_NO_NEW_PRIVS type (i32 not u64)
- Fix sync_fd handling in process.rs (RawFd vs OwnedFd)
- Fix mount() call in storage.rs to use CString for options
- Fix storage.rs unmount to use blocking RwLock instead of async

+++ crates/exo-runtime/src/agent.rs
@@ -364,7 +364,7 @@ impl AgentProfile {
     #[cfg(target_os = "linux")]
     fn set_no_new_privs(&self) -> Result<()> {
         use nix::unistd::Uid;
-        const PR_SET_NO_NEW_PRIVS: u64 = 38;
+        const PR_SET_NO_NEW_PRIVS: i32 = 38;
 
         // Use prctl to set no_new_privs
         let ret = unsafe { libc::prctl(PR_SET_NO_NEW_PRIVS, 1, 0, 0, 0) };
